Changes Made: - Added the aggregatedTruncTotal Measure and the absAggregatedTruncTotal Measure. - Added the tests for these Measures. - Made amendments to the function to not include double casts. - Changed the result from BigDecimal to LongValue.

+++ agent/src/main/scala/za/co/absa/atum/agent/model/Measure.scala
@@ -39,16 +39,6 @@ final case class UnknownMeasure(measureName: String, measuredColumns: Seq[String
 
 object AtumMeasure {
 
-//  val supportedMeasureNames: Seq[String] = Seq(
-//    RecordCount.measureName,
-//    DistinctRecordCount.measureName,
-//    SumOfValuesOfColumn.measureName,
-//    AbsSumOfValuesOfColumn.measureName,
-//    SumOfTruncatedValuesOfColumn.measureName,
-//    AbsSumOfTruncatedValuesOfColumn.measureName,
-//    SumOfHashesOfColumn.measureName
-//  )
-
   case class RecordCount private (measureName: String) extends AtumMeasure {
     private val columnExpression = count("*")
 
@@ -120,8 +110,7 @@ object AtumMeasure {
   }
 
   case class SumOfTruncatedValuesOfColumn private (measureName: String, measuredCol: String) extends AtumMeasure {
-    //Cast to LongType to remove decimal points then cast back to decimal to ensure compatibility
-    //private val columnAggFn: Column => Column = column => sum(column.cast(LongType).cast(DecimalType(38, 0)))
+
     private val columnAggFn: Column => Column = column => sum(when(column >= 0, floor(column)).otherwise(ceil(column)))
 
     override def function: MeasurementFunction = (ds: DataFrame) => {
@@ -131,7 +120,7 @@ object AtumMeasure {
     }
 
     override def measuredColumns: Seq[String] = Seq(measuredCol)
-    override val resultValueType: ResultValueType = ResultValueType.BigDecimalValue
+    override val resultValueType: ResultValueType = ResultValueType.LongValue
   }
   object SumOfTruncatedValuesOfColumn {
     private[agent] val measureName: String = "aggregatedTruncTotal"
@@ -139,8 +128,7 @@ object AtumMeasure {
   }
 
   case class AbsSumOfTruncatedValuesOfColumn private (measureName: String, measuredCol: String) extends AtumMeasure {
-    //Cast to LongType to remove decimal points then cast back to decimal to ensure compatibility
-    //private val columnAggFn: Column => Column = column => sum(abs(column.cast(LongType).cast(DecimalType(38, 0))))
+
     private val columnAggFn: Column => Column = column => sum(abs(when(column >= 0, floor(column)).otherwise(ceil(column))))
 
     override def function: MeasurementFunction = (ds: DataFrame) => {
@@ -150,7 +138,7 @@ object AtumMeasure {
     }
 
     override def measuredColumns: Seq[String] = Seq(measuredCol)
-    override val resultValueType: ResultValueType = ResultValueType.BigDecimalValue
+    override val resultValueType: ResultValueType = ResultValueType.LongValue
   }
   object AbsSumOfTruncatedValuesOfColumn {
     private[agent] val measureName: String = "absAggregatedTruncTotal"
