+++ agent/src/test/scala/za/co/absa/atum/agent/model/AtumMeasureUnitTests.scala
@@ -129,13 +129,13 @@ class AtumMeasureUnitTests extends AnyFlatSpec with Matchers with SparkTestBase 
     assert(dfFullSalarySumResult.resultValue == "2987144")
     assert(dfFullSalarySumResult.resultValueType == ResultValueType.BigDecimalValue)
     assert(dfExtraPersonSalarySumTruncResult.resultValue == "2989144")
-    assert(dfExtraPersonSalarySumTruncResult.resultValueType == ResultValueType.BigDecimalValue)
+    assert(dfExtraPersonSalarySumTruncResult.resultValueType == ResultValueType.LongValue)
     assert(dfFullSalarySumTruncResult.resultValue == "2987144")
-    assert(dfFullSalarySumTruncResult.resultValueType == ResultValueType.BigDecimalValue)
+    assert(dfFullSalarySumTruncResult.resultValueType == ResultValueType.LongValue)
     assert(dfExtraPersonSalaryAbsSumTruncResult.resultValue == "2991144")
-    assert(dfExtraPersonSalaryAbsSumTruncResult.resultValueType == ResultValueType.BigDecimalValue)
+    assert(dfExtraPersonSalaryAbsSumTruncResult.resultValueType == ResultValueType.LongValue)
     assert(dfFullSalaryAbsSumTruncResult.resultValue == "2987144")
-    assert(dfFullSalaryAbsSumTruncResult.resultValueType == ResultValueType.BigDecimalValue)
+    assert(dfFullSalaryAbsSumTruncResult.resultValueType == ResultValueType.LongValue)
   }
 
   "AbsSumOfValuesOfColumn" should "return expected value" in {
@@ -229,7 +229,7 @@ class AtumMeasureUnitTests extends AnyFlatSpec with Matchers with SparkTestBase 
     val result = distinctCount.function(df)
 
     assert(result.resultValue == "2")
-    assert(result.resultValueType == ResultValueType.BigDecimalValue)
+    assert(result.resultValueType == ResultValueType.LongValue)
   }
 
   "AbsSumTruncOfValuesOfColumn" should "return expected value" in {
@@ -244,6 +244,6 @@ class AtumMeasureUnitTests extends AnyFlatSpec with Matchers with SparkTestBase 
     val result = distinctCount.function(df)
 
     assert(result.resultValue == "4")
-    assert(result.resultValueType == ResultValueType.BigDecimalValue)
+    assert(result.resultValueType == ResultValueType.LongValue)
   }
 }
